anomalydetector/requirements.txt
quixstreams<2.10
python-dotenv
scikit-learn